% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{apothecary_explicit_model}
\alias{apothecary_explicit_model}
\title{Apothecary model creation.

We will use this structure to ensure that model fitting is flexible in the
future as more models are added}
\usage{
apothecary_explicit_model()
}
\description{
Create an apothecary model
}
