% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta_est_apothecary}
\alias{beta_est_apothecary}
\title{Estimate beta parameter for apothecary model}
\usage{
beta_est_apothecary(
  dur_IAsymp,
  dur_IMild,
  dur_ICase,
  dur_IPreAsymp,
  dur_IPreMild,
  dur_IPreCase,
  rel_inf_asymp,
  rel_inf_mild,
  prob_asymp,
  prob_hosp,
  mixing_matrix,
  R0
)
}
\arguments{
\item{dur_IAsymp}{Duration of asymptomatic infectiousness (days)}

\item{dur_IMild}{Duration of mild infectiousness (days)}

\item{dur_ICase}{Delay between symptom onset and requiring hospitalisation (days)}

\item{rel_inf_asymp}{Relative infectiousness of asymptomatics relative to ICase}

\item{rel_inf_mild}{Relative infeciousness of mild infections relative to ICase}

\item{prob_asymp}{Probability of asymptomatic infection by age}

\item{prob_hosp}{Probability of hospitilisation by age}

\item{mixing_matrix}{Mixing matrix}
}
\value{
Eigenvalue
}
\description{
Estimate beta parameter for apothecary model
}
