% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{default_durations}
\alias{default_durations}
\title{Return the default hospital durations for modelling}
\usage{
default_durations()
}
\value{
List of default durations
}
\description{
Return the default hospital durations for modelling
}
