% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{div_pop}
\alias{div_pop}
\title{Divide matrix by population}
\usage{
div_pop(contact, population)
}
\arguments{
\item{contact}{Matrix}

\item{population}{Population vector}
}
\value{
Matrix
}
\description{
Divide matrix by population
}
