% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{initial_params}
\alias{initial_params}
\title{Generate initial parameters for the apothecary model}
\usage{
initial_params(country = NULL, population = NULL, seeding_cases = 20)
}
\arguments{
\item{country}{country}

\item{population}{population}

\item{seeding_cases}{seeding_cases}
}
\value{
inits
}
\description{
Generate initial parameters for the apothecary model
}
