% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_apothecary}
\alias{run_apothecary}
\title{Run the apothecary SEIR model}
\usage{
run_apothecary(
  model = "deterministic",
  country = NULL,
  population = NULL,
  tt_contact_matrix = 0,
  contact_matrix_set = NULL,
  hosp_bed_capacity = NULL,
  ICU_bed_capacity = NULL,
  tt_hosp_beds = 0,
  tt_ICU_beds = 0,
  prop_ox_hosp_beds = 1,
  prop_ox_ICU_beds = 1,
  tt_prop_ox_hosp_beds = 0,
  tt_prop_ox_ICU_beds = 0,
  MV_capacity = 1e+09,
  R0 = 3,
  tt_R0 = 0,
  beta_set = NULL,
  rel_inf_asymp = 1,
  rel_inf_mild = 1,
  tt_matrix = 0,
  dt = NULL,
  N_age = 17,
  time_period = 365,
  init = NULL,
  seeding_cases = NULL,
  prob_asymp = NULL,
  prob_hosp = NULL,
  prob_severe = NULL,
  prob_critical = NULL,
  prob_moderate_death_get_hosp_get_ox_baseline = NULL,
  prob_moderate_death_get_hosp_no_ox_baseline = NULL,
  prob_moderate_death_no_hosp_no_ox = NULL,
  prob_severe_death_get_ICU_get_ox_baseline = NULL,
  prob_severe_death_get_ICU_no_ox_baseline = NULL,
  prob_severe_death_no_ICU_no_ox = NULL,
  prob_critical_death_get_ICU_get_ox_get_MV_baseline = NULL,
  prob_critical_death_get_ICU_get_ox_no_MV_baseline = NULL,
  prob_critical_death_get_ICU_no_ox_no_MV_baseline = NULL,
  prob_critical_death_no_ICU_no_ox_no_MV = NULL,
  dur_E = NULL,
  dur_IPreAsymp = NULL,
  dur_IPreMild = NULL,
  dur_IPreCase = NULL,
  dur_IAsymp = NULL,
  dur_IMild = NULL,
  dur_ICase = NULL,
  dur_rec = NULL,
  dur_IMod_GetHosp_GetOx_Surv = NULL,
  dur_IMod_GetHosp_GetOx_Die = NULL,
  dur_IMod_GetHosp_NoOx_Surv = NULL,
  dur_IMod_GetHosp_NoOx_Die = NULL,
  dur_IMod_NoHosp_NoOx_Surv = NULL,
  dur_IMod_NoHosp_NoOx_Die = NULL,
  dur_ISev_GetICU_GetOx_Surv = NULL,
  dur_ISev_GetICU_GetOx_Die = NULL,
  dur_ISev_GetICU_NoOx_Surv = NULL,
  dur_ISev_GetICU_NoOx_Die = NULL,
  dur_ISev_NoICU_NoOx_Surv = NULL,
  dur_ISev_NoICU_NoOx_Die = NULL,
  dur_ICrit_GetICU_GetOx_GetMV_Surv = NULL,
  dur_ICrit_GetICU_GetOx_GetMV_Die = NULL,
  dur_ICrit_GetICU_GetOx_NoMV_Surv = NULL,
  dur_ICrit_GetICU_GetOx_NoMV_Die = NULL,
  dur_ICrit_GetICU_NoOx_NoMV_Surv = NULL,
  dur_ICrit_GetICU_NoOx_NoMV_Die = NULL,
  dur_ICrit_NoICU_NoOx_NoMV_Surv = NULL,
  dur_ICrit_NoICU_NoOx_NoMV_Die = NULL,
  prophylactic_drug_timing_1 = 1e+05,
  prophylactic_drug_timing_2 = 1e+05,
  prophylactic_prop_treat = 0,
  prophylactic_drug_wane = 1e+06,
  drug_1_indic = 0,
  drug_1_effect_size = 0,
  drug_2_indic_IPreAsymp = 0,
  drug_2_indic_IPreMild = 0,
  drug_2_indic_IPreCase = 0,
  drug_2_prop_treat = 0,
  drug_2_effect_size_IPreAsymp = 1,
  drug_2_effect_size_IPreMild = 1,
  drug_2_effect_size_IPreCase = 1,
  drug_3_indic = 0,
  drug_3_prop_treat = 0,
  drug_3_effect_size = 1,
  drug_4_indic_IAsymp = 0,
  drug_4_indic_IMild = 0,
  drug_4_indic_ICase = 0,
  drug_4_prop_treat = 0,
  drug_4_effect_size_IAsymp = 1,
  drug_4_effect_size_IMild = 1,
  drug_4_effect_size_ICase = 1,
  drug_5_indic_IMild = 0,
  drug_5_indic_ICase = 0,
  drug_5_prop_treat = 0,
  drug_5_effect_size = 0,
  drug_6_indic = 0,
  drug_6_prop_treat = 0,
  drug_6_effect_size = 0,
  drug_7_indic = 0,
  drug_7_prop_treat = 0,
  drug_7_effect_size = 0,
  drug_8_indic_IMod_GetHosp_GetOx = 0,
  drug_8_indic_IMod_GetHosp_NoOx = 0,
  drug_8_prop_treat = 0,
  drug_8_GetOx_effect_size = 1,
  drug_8_NoOx_effect_size = 1,
  drug_9_indic_ISev_GetICU_GetOx = 0,
  drug_9_indic_ISev_GetICU_NoOx = 0,
  drug_9_prop_treat = 0,
  drug_9_GetOx_effect_size = 1,
  drug_9_NoOx_effect_size = 1,
  drug_10_indic_ICrit_GetICU_GetOx_GetMV = 0,
  drug_10_indic_ICrit_GetICU_GetOx_NoMV = 0,
  drug_10_indic_ICrit_GetICU_NoOx_NoMV = 0,
  drug_10_prop_treat = 0,
  drug_10_GetOx_GetMV_effect_size = 1,
  drug_10_GetOx_NoMV_effect_size = 1,
  drug_10_NoOx_NoMV_effect_size = 1,
  drug_11_indic_IMod_GetHosp_GetOx = 0,
  drug_11_indic_IMod_GetHosp_NoOx = 0,
  drug_11_prop_treat = 0,
  drug_11_GetOx_effect_size = 1,
  drug_11_NoOx_effect_size = 1,
  drug_12_indic_ISev_GetICU_GetOx = 0,
  drug_12_indic_ISev_GetICU_NoOx = 0,
  drug_12_prop_treat = 0,
  drug_12_GetOx_effect_size = 1,
  drug_12_NoOx_effect_size = 1,
  drug_13_indic_ICrit_GetICU_GetOx_GetMV = 0,
  drug_13_indic_ICrit_GetICU_GetOx_NoMV = 0,
  drug_13_indic_ICrit_GetICU_NoOx_NoMV = 0,
  drug_13_prop_treat = 0,
  drug_13_GetOx_GetMV_effect_size = 1,
  drug_13_GetOx_NoMV_effect_size = 1,
  drug_13_NoOx_NoMV_effect_size = 1,
  day_return = TRUE,
  replicates = 1,
  walker_params = FALSE
)
}
\description{
Run the apothecary SEIR model
}
