% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_apothecary}
\alias{run_apothecary}
\title{Run the apothecary SEIR model}
\usage{
run_apothecary(
  model = "deterministic",
  country = NULL,
  population = NULL,
  tt_contact_matrix = 0,
  contact_matrix_set = NULL,
  hosp_bed_capacity = NULL,
  ICU_bed_capacity = NULL,
  tt_hosp_beds = 0,
  tt_ICU_beds = 0,
  tt_oxygen_supply = 0,
  input_oxygen_supply = 1e+09,
  tt_baseline_oxygen_demand = 0,
  input_baseline_oxygen_demand = 0,
  oxygen_availability_0 = 1e+09,
  max_leftover = 0,
  severe_critical_case_oxygen_consumption_multiplier = 1.5,
  MV_capacity = 1e+09,
  R0 = 3,
  tt_R0 = 0,
  beta_set = NULL,
  rel_inf_asymp = 1,
  rel_inf_mild = 1,
  tt_matrix = 0,
  dt = 0.1,
  N_age = 17,
  time_period = 365,
  init = NULL,
  seeding_cases = NULL,
  prob_asymp = rep(0.5, 17),
  prob_hosp = c(0.000744192, 0.000634166, 0.001171109, 0.002394593, 0.005346437,
    0.010289885, 0.016234604, 0.023349169, 0.028944623, 0.038607042, 0.057734879,
    0.072422135, 0.101602458, 0.116979814, 0.146099064, 0.176634654, 0.18),
  prob_severe = c(0.05022296, 0.05022296, 0.05022296, 0.05022296, 0.05022296,
    0.05022296, 0.05022296, 0.053214942, 0.05974426, 0.074602879, 0.103612417,
    0.149427991, 0.223777304, 0.306985918, 0.385779555, 0.461217861, 0.709444444),
  prob_critical = rep(0.8, 17),
  prob_moderate_death_get_hosp_get_ox_baseline = c(0.0125702, 0.0125702, 0.0125702,
    0.0125702, 0.0125702, 0.0125702, 0.0125702, 0.013361147, 0.015104687, 0.019164124,
    0.027477519, 0.041762108, 0.068531658, 0.105302319, 0.149305732, 0.20349534,
    0.5804312),
  prob_moderate_death_get_hosp_no_ox_baseline = rep(0.6, 17),
  prob_moderate_death_no_hosp_no_ox = rep(0.6, 17),
  prob_severe_death_get_ICU_get_ox_baseline = rep(0.5, 17),
  prob_severe_death_get_ICU_no_ox_baseline = rep(0.95, 17),
  prob_severe_death_no_ICU_no_ox = rep(0.95, 17),
  prob_critical_death_get_ICU_get_ox_get_MV_baseline = rep(0.5, 17),
  prob_critical_death_get_ICU_get_ox_no_MV_baseline = rep(0.95, 17),
  prob_critical_death_get_ICU_no_ox_no_MV_baseline = rep(0.95, 17),
  prob_critical_death_no_ICU_no_ox_no_MV = rep(0.95, 17),
  dur_E = 4.6,
  dur_IAsymp = 2.1,
  dur_IMild = 2.1,
  dur_ICase = 4.5,
  dur_rec = 3.4,
  dur_IMod_GetHosp_GetOx_Surv = 9.6,
  dur_IMod_GetHosp_GetOx_Die = 7.6,
  dur_IMod_GetHosp_NoOx_Surv = 9.6/2,
  dur_IMod_GetHosp_NoOx_Die = 7.6/2,
  dur_IMod_NoHosp_NoOx_Surv = 9.6/2,
  dur_IMod_NoHosp_NoOx_Die = 7.6/2,
  dur_ISev_GetICU_GetOx_Surv = 11.3,
  dur_ISev_GetICU_GetOx_Die = 10.1,
  dur_ISev_GetICU_NoOx_Surv = 11.3/2,
  dur_ISev_GetICU_NoOx_Die = 10.1/2,
  dur_ISev_NoICU_NoOx_Surv = 11.3/2,
  dur_ISev_NoICU_NoOx_Die = 10.1/2,
  dur_ICrit_GetICU_GetOx_GetMV_Surv = 11.3,
  dur_ICrit_GetICU_GetOx_GetMV_Die = 10.1,
  dur_ICrit_GetICU_GetOx_NoMV_Surv = 11.3/2,
  dur_ICrit_GetICU_GetOx_NoMV_Die = 10.1/2,
  dur_ICrit_GetICU_NoOx_NoMV_Surv = 11.3/2,
  dur_ICrit_GetICU_NoOx_NoMV_Die = 10.1/2,
  dur_ICrit_NoICU_NoOx_NoMV_Surv = 11.3/2,
  dur_ICrit_NoICU_NoOx_NoMV_Die = 10.1/2,
  prophylactic_drug_timing_1 = 1e+05,
  prophylactic_drug_timing_2 = 1e+05,
  prophylactic_prop_treat = 0,
  prophylactic_drug_wane = 1e+06,
  drug_1_indic = 0,
  drug_1_effect_size = 0,
  drug_2_indic = 0,
  drug_2_effect_size = 0,
  drug_3_indic = 0,
  drug_3_prop_treat = 0,
  drug_3_effect_size = 0,
  drug_4_indic = 0,
  drug_4_prop_treat = 0,
  drug_4_effect_size = 0,
  drug_5_indic_IMild = 0,
  drug_5_indic_ICase = 0,
  drug_5_prop_treat = 0,
  drug_5_effect_size = 0,
  drug_6_indic = 0,
  drug_6_prop_treat = 0,
  drug_6_effect_size = 0,
  drug_7_indic = 0,
  drug_7_prop_treat = 0,
  drug_7_effect_size = 0,
  drug_8_indic_IMod_GetHosp_GetOx = 0,
  drug_8_indic_IMod_GetHosp_NoOx = 0,
  drug_8_prop_treat = 0,
  drug_8_GetOx_effect_size = 0,
  drug_8_NoOx_effect_size = 0,
  drug_9_indic_ISev_GetICU_GetOx = 0,
  drug_9_indic_ISev_GetICU_NoOx = 0,
  drug_9_prop_treat = 0,
  drug_9_GetOx_effect_size = 0,
  drug_9_NoOx_effect_size = 0,
  drug_10_indic_ICrit_GetICU_GetOx_GetMV = 0,
  drug_10_indic_ICrit_GetICU_GetOx_NoMV = 0,
  drug_10_indic_ICrit_GetICU_NoOx_NoMV = 0,
  drug_10_prop_treat = 0,
  drug_10_GetOx_GetMV_effect_size = 0,
  drug_10_GetOx_NoMV_effect_size = 0,
  drug_10_NoOx_NoMV_effect_size = 0,
  drug_11_indic_IMod_GetHosp_GetOx = 0,
  drug_11_indic_IMod_GetHosp_NoOx = 0,
  drug_11_prop_treat = 0,
  drug_11_GetOx_effect_size = 0,
  drug_11_NoOx_effect_size = 0,
  drug_12_indic_ISev_GetICU_GetOx = 0,
  drug_12_indic_ISev_GetICU_NoOx = 0,
  drug_12_prop_treat = 0,
  drug_12_GetOx_effect_size = 0,
  drug_12_NoOx_effect_size = 0,
  drug_13_indic_ICrit_GetICU_GetOx_GetMV = 0,
  drug_13_indic_ICrit_GetICU_GetOx_NoMV = 0,
  drug_13_indic_ICrit_GetICU_NoOx_NoMV = 0,
  drug_13_prop_treat = 0,
  drug_13_GetOx_GetMV_effect_size = 0,
  drug_13_GetOx_NoMV_effect_size = 0,
  drug_13_NoOx_NoMV_effect_size = 0,
  day_return = TRUE,
  replicates = 1
)
}
\description{
Run the apothecary SEIR model
}
